pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baymax"
version = "0.1.0"
description = "BayMax Stock Agent - An AI-powered financial research agent for stock analysis and investment research."
authors = [{name = "BayMax Stock Agent Team", email = "contact@baymax-stock-agent.com"}]
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langchain-anthropic>=0.3.1",
    "langchain-google-genai>=2.0.0",
    "openai>=2.2.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "akshare>=1.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "httpx[socks]>=0.28.1",
]

[project.scripts]
baymax = "baymax.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
